requirements.txt
anyio==3.7.1
python-dotenv==0.21.1
python-telegram-bot==20.3